chromium_ec: Extract i2c_passthrough to separate module
TODO: Split out port and address code also
